Abstract Node used for Network diagram
