+++ apps/website/src/components/tech/Node.tsx
@@ -1,0 +1,99 @@
+import { useMemo } from "react";
+import { Handle, type NodeProps, Position, useEdges } from "reactflow";
+
+import { classes } from "@/utils/classes";
+
+import IconExternal from "@/icons/IconExternal";
+
+export interface NodeData {
+  item: {
+    container: string;
+    eyebrow: { text: string; color: string };
+    name: string;
+    description?: string;
+    url?: string;
+  };
+}
+
+const Node = ({
+  id,
+  data,
+  targetPosition = Position.Top,
+  sourcePosition = Position.Bottom,
+  isConnectable,
+}: NodeProps<NodeData>) => {
+  // Get the source and target edges
+  const edges = useEdges();
+  let targetEdge, sourceEdge;
+  for (const edge of edges) {
+    if (!targetEdge && edge.target === id) targetEdge = edge;
+    if (!sourceEdge && edge.source === id) sourceEdge = edge;
+    if (targetEdge && sourceEdge) break;
+  }
+
+  // If this node has a link, we need some extra props
+  const Element = data.item.url ? "a" : "div";
+  const linkProps = useMemo(
+    () =>
+      data.item.url
+        ? {
+            href: data.item.url,
+            target: "_blank",
+            rel: "noopener noreferrer",
+          }
+        : {},
+    [data.item.url],
+  );
+
+  return (
+    <Element
+      className={classes(
+        "group flex h-20 w-44 cursor-pointer flex-col rounded-xl border-2 bg-white px-2 py-1 hover:min-w-min hover:shadow-md focus:min-w-min focus:shadow-md",
+        data.item.container,
+      )}
+      tabIndex={-1}
+      {...linkProps}
+    >
+      {(targetEdge || isConnectable) && (
+        <Handle
+          type="target"
+          position={targetPosition}
+          isConnectable={isConnectable}
+        />
+      )}
+      {(sourceEdge || isConnectable) && (
+        <Handle
+          type="source"
+          position={sourcePosition}
+          isConnectable={isConnectable}
+        />
+      )}
+
+      <p className={classes("text-xs", data.item.eyebrow.color)}>
+        {data.item.eyebrow.text}
+      </p>
+      <div className="my-auto">
+        {data.item.description && (
+          <p className="truncate text-alveus-green-900">{data.item.name}</p>
+        )}
+
+        <p className="flex items-center gap-1 text-xs text-alveus-green-700">
+          <span
+            className={classes(
+              "shrink overflow-hidden text-ellipsis whitespace-nowrap",
+              data.item.url && "group-hover:underline",
+            )}
+          >
+            {data.item.description || data.item.name}
+          </span>
+
+          {data.item.url && (
+            <IconExternal className="shrink-0 grow-0" size={14} />
+          )}
+        </p>
+      </div>
+    </Element>
+  );
+};
+
+export default Node;
